backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
bitarray==2.9.2
bitstring==4.2.3
cffi==1.17.1
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
esptool==4.8.0
fastapi==0.115.0
greenlet==3.1.0
h11==0.14.0
idna==3.10
intelhex==2.3.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
httpx==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
